- Utility classes reorganization.
