+++ .env.dist
@@ -1,3 +1,3 @@
-VERSION=1.7.4
+VERSION=1.8.0
 PYPI_USERNAME=__token__
 PYPI_PASSWORD= #api token
