+++ pytermor/common.py
