+++ pytermor/numf/__init__.py
@@ -2,29 +2,18 @@
 # pytermor [ANSI formatted terminal output toolset]
 # (C) 2022 A. Shavykin <0.delameter@gmail.com>
 # -----------------------------------------------------------------------------
-from .string_filter import *
-from .fmtd import *
-
 from .auto_float import *
 from .prefixed_unit import *
 from .time_delta import *
 
 __all__ = [
-    'apply_filters',
-    'StringFilter',
-    'ReplaceCSI',
-    'ReplaceSGR',
-    'ReplaceNonAsciiBytes',
+    'format_auto_float',
 
-    'ljust_fmtd',
-    'rjust_fmtd',
-    'center_fmtd',
-
-    'fmt_prefixed_unit',
+    'format_prefixed_unit',
     'PrefixedUnitPreset',
-    'FMT_PRESET_SI_METRIC',
-    'FMT_PRESET_SI_BINARY',
-    'fmt_time_delta',
+    'PRESET_SI_METRIC',
+    'PRESET_SI_BINARY',
+
+    'format_time_delta',
     'TimeDeltaPreset',
-    'fmt_auto_float',
 ]
