+++ pytermor/strf/__init__.py
@@ -2,20 +2,10 @@
 # pytermor [ANSI formatted terminal output toolset]
 # (C) 2022 A. Shavykin <0.delameter@gmail.com>
 # -----------------------------------------------------------------------------
-from .seq import build, build_c256, build_rgb, SequenceSGR
-from .fmt import autof, Format
-from .numf import *
-from .strf import *
+from .string_filter import *
+from .fmtd import *
 
 __all__ = [
-    'build',
-    'build_c256',
-    'build_rgb',
-    'SequenceSGR',
-
-    'autof',
-    'Format',
-
     'apply_filters',
     'StringFilter',
     'ReplaceCSI',
@@ -25,13 +15,4 @@ __all__ = [
     'ljust_fmtd',
     'rjust_fmtd',
     'center_fmtd',
-
-    'format_auto_float',
-    'format_prefixed_unit',
-    'PrefixedUnitPreset',
-    'PRESET_SI_METRIC',
-    'PRESET_SI_BINARY',
-    'format_time_delta',
-    'TimeDeltaPreset',
 ]
-__version__ = '1.8.0'
