+++ pytermor/strf/fmtd.py
