+++ pytermor/strf/string_filter.py
