+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = pytermor
-version = 1.7.4
+version = 1.8.0
 author = Aleksandr Shavykin
 author_email = 0.delameter@gmail.com
 description = ANSI formatted terminal output library
