+++ tests/__init__.py
@@ -1,0 +1,6 @@
+from .common import *
+
+__all__ = [
+    'is_verbose_mode',
+    'print_verbose',
+]
