+++ tests/numf/test_time_delta.py
@@ -1,0 +1,105 @@
+# -----------------------------------------------------------------------------
+# pytermor [ANSI formatted terminal output toolset]
+# (C) 2022 A. Shavykin <0.delameter@gmail.com>
+# -----------------------------------------------------------------------------
+import unittest
+from datetime import timedelta
+
+from pytermor import format_time_delta
+from tests import print_verbose
+
+
+class TestTimeDelta(unittest.TestCase):
+    expected_format_max_len = 10
+    expected_format_dataset = [
+        ['OVERFLOW', timedelta(days=-700000)],
+        ['-2 years', timedelta(days=-1000)],
+        ['-10 months', timedelta(days=-300)],
+        ['-3 months', timedelta(days=-100)],
+        ['-9d 23h', timedelta(days=-9, hours=-23)],
+        ['-5d 0h', timedelta(days=-5)],
+        ['-13h 30min', timedelta(days=-1, hours=10, minutes=30)],
+        ['-45 mins', timedelta(hours=-1, minutes=15)],
+        ['-5 mins', timedelta(minutes=-5)],
+        ['-2 secs', timedelta(seconds=-2.01)],
+        ['-2 secs', timedelta(seconds=-2)],
+        ['-1 sec', timedelta(seconds=-2, microseconds=1)],
+        ['-1 sec', timedelta(seconds=-1.9)],
+        ['-1 sec', timedelta(seconds=-1.1)],
+        ['-1 sec', timedelta(seconds=-1.0)],
+        ['~0 secs', timedelta(seconds=-0.5)],
+        ['~0 secs', timedelta(milliseconds=-50)],
+        ['~0 secs', timedelta(microseconds=-100)],
+        ['~0 secs', timedelta(microseconds=-1)],
+        ['0 secs', timedelta()],
+        ['0 secs', timedelta(microseconds=500)],
+        ['<1 sec', timedelta(milliseconds=25)],
+        ['<1 sec', timedelta(seconds=0.1)],
+        ['<1 sec', timedelta(seconds=0.9)],
+        ['1 sec', timedelta(seconds=1)],
+        ['1 sec', timedelta(seconds=1.0)],
+        ['1 sec', timedelta(seconds=1.1)],
+        ['1 sec', timedelta(seconds=1.9)],
+        ['1 sec', timedelta(seconds=2, microseconds=-1)],
+        ['2 secs', timedelta(seconds=2)],
+        ['2 secs', timedelta(seconds=2.0)],
+        ['2 secs', timedelta(seconds=2.5)],
+        ['10 secs', timedelta(seconds=10)],
+        ['1 min', timedelta(minutes=1)],
+        ['5 mins', timedelta(minutes=5)],
+        ['15 mins', timedelta(minutes=15)],
+        ['45 mins', timedelta(minutes=45)],
+        ['1h 30min', timedelta(hours=1, minutes=30)],
+        ['4h 15min', timedelta(hours=4, minutes=15)],
+        ['8h 59min', timedelta(hours=8, minutes=59, seconds=59)],
+        ['12h 30min', timedelta(hours=12, minutes=30)],
+        ['18h 45min', timedelta(hours=18, minutes=45)],
+        ['23h 50min', timedelta(hours=23, minutes=50)],
+        ['1d 0h', timedelta(days=1)],
+        ['3d 4h', timedelta(days=3, hours=4)],
+        ['5d 22h', timedelta(days=5, hours=22, minutes=51)],
+        ['6d 23h', timedelta(days=7, minutes=-1)],
+        ['9d 0h', timedelta(days=9)],
+        ['12 days', timedelta(days=12, hours=18)],
+        ['16 days', timedelta(days=16, hours=2)],
+        ['1 month', timedelta(days=30)],
+        ['1 month', timedelta(days=55)],
+        ['2 months', timedelta(days=70)],
+        ['2 months', timedelta(days=80)],
+        ['6 months', timedelta(days=200)],
+        ['11 months', timedelta(days=350)],
+        ['1 year', timedelta(days=390)],
+        ['2 years', timedelta(days=810)],
+        ['27 years', timedelta(days=10000)],
+        ['277 years', timedelta(days=100000)],
+        ['OVERFLOW', timedelta(days=400000)],
+    ]
+
+    def test_output_has_expected_format(self):
+        for idx, (expected_output, input_arg) in enumerate(self.expected_format_dataset):
+            with self.subTest(msg=f'Test #{idx}: "{input_arg}" -> "{expected_output}"'):
+                if type(input_arg) is not timedelta:
+                    input_arg = timedelta(days=input_arg)
+
+                actual_output = format_time_delta(input_arg.total_seconds(), self.expected_format_max_len)
+                print_verbose(actual_output, end=', ')
+                self.assertEqual(expected_output, actual_output)
+
+    req_len_expected_len_list = [3, 4, 6, 10, 9, 1000]
+    req_len_input_delta_list = [el[1] for el in expected_format_dataset]
+
+    def test_output_fits_in_required_length(self):
+        for idx, expected_max_len in enumerate(self.req_len_expected_len_list):
+            print_verbose(f'\nLEN={expected_max_len:d}', end=': ')
+            for input_td in self.req_len_input_delta_list:
+                with self.subTest(msg=f'Testing #{idx} "{input_td}" -> max length {expected_max_len}'):
+                    actual_output = format_time_delta(input_td.total_seconds(), expected_max_len)
+                    print_verbose(actual_output, end=', ')
+                    self.assertGreaterEqual(expected_max_len, len(actual_output))
+
+    invalid_len_list = [-5, 0, 1, 2]
+
+    def test_invalid_max_length_fails(self):
+        for invalid_max_len in self.invalid_len_list:
+            with self.subTest(msg=f'Testing invalid max length {invalid_max_len}'):
+                self.assertRaises(ValueError, lambda: format_time_delta(100, invalid_max_len))
