+++ tests/strf/__init__.py
@@ -1,0 +1,13 @@
+# -----------------------------------------------------------------------------
+# pytermor [ANSI formatted terminal output toolset]
+# (C) 2022 A. Shavykin <0.delameter@gmail.com>
+# -----------------------------------------------------------------------------
+import unittest
+
+
+class TestStringFilter(unittest.TestCase):
+    pass  # @TODO
+
+
+class TestFmtd(unittest.TestCase):
+    pass  # @TODO
